+++ src/py_eddy_tracker/observations/observation.py
@@ -1614,7 +1614,7 @@ class EddiesObservations(object):
         :param matplotlib.axes.Axes ax: matplotlib axe used to draw
         :param str,array,None name:
             variable used to fill the contour, if None all elements have the same color
-        :param float,None ref: if define use like west bound ?
+        :param float,None ref: if define use like west bound
         :param float factor: multiply value by
         :param dict kwargs: look at :py:meth:`matplotlib.axes.Axes.scatter`
         :return: scatter mappable